Split galley API routes
