+++ libs/wire-api-federation/test/Test/Wire/API/Federation/Golden/MessageSendResponse.hs
@@ -25,7 +25,7 @@ import GHC.Exts (IsList (fromList))
 import Imports
 import Wire.API.Federation.API.Galley (MessageSendResponse (..))
 import Wire.API.Message
-import Wire.API.Routes.Public.Galley
+import Wire.API.Routes.Public.Galley.Messaging
 import Wire.API.User.Client (QualifiedUserClients (..))
 
 missing :: QualifiedUserClients
