+++ libs/wire-api/wire-api.cabal
@@ -91,6 +91,16 @@ library
     Wire.API.Routes.Public.Cannon
     Wire.API.Routes.Public.Cargohold
     Wire.API.Routes.Public.Galley
+    Wire.API.Routes.Public.Galley.Bot
+    Wire.API.Routes.Public.Galley.Conversation
+    Wire.API.Routes.Public.Galley.CustomBackend
+    Wire.API.Routes.Public.Galley.Feature
+    Wire.API.Routes.Public.Galley.LegalHold
+    Wire.API.Routes.Public.Galley.Messaging
+    Wire.API.Routes.Public.Galley.MLS
+    Wire.API.Routes.Public.Galley.Team
+    Wire.API.Routes.Public.Galley.TeamConversation
+    Wire.API.Routes.Public.Galley.TeamMember
     Wire.API.Routes.Public.Gundeck
     Wire.API.Routes.Public.Spar
     Wire.API.Routes.Public.Util
